backend/requirements.txt
matplotlib
torch
bs4
pandas
tqdm
transformers
numpy
kafka-python

Flask
jupyter
flask-cors
pymysql
google-cloud-pubsub
grpcio
google-auth
firebase_admin
exponent_server_sdk